Support zero-size `Mmap` and `MmapMut`
Fixes https://github.com/danburkert/memmap-rs/issues/72

Previously, mapping an empty file or requesting a zero-size mapping would
return an error because underlying OS APIs do not support zero-size mappings.
Now, this crate instead requests a one-byte mapping (which most lilkely ends up
rounded up to one memory page) but still has `Deref` return an empty slice
in those cases.

This is the behavior suggested for Unix in
https://github.com/danburkert/memmap-rs/issues/72#issuecomment-428872242

On Windows, docs for `CreateFileMappingW` say:
https://docs.microsoft.com/en-us/windows/win32/api/memoryapi/nf-memoryapi-createfilemappingw
> `dwMaximumSizeHigh`
>
> The high-order DWORD of the maximum size of the file mapping object.
>
> `dwMaximumSizeLow`
>
> The low-order DWORD of the maximum size of the file mapping object.
>
> If this parameter and dwMaximumSizeHigh are 0 (zero), the maximum size of
> the file mapping object is equal to the current size of the file that hFile
> identifies.
>
> An attempt to map a file with a length of 0 (zero) fails with an error code
> of ERROR_FILE_INVALID. Applications should test for files with a length of
> 0 (zero) and reject those files.

Hopefully that last paragraph only applies when `dwMaximumSize` is zero
(which this PR avoids), but that is not clear to me.