Fix mapping empty files on Windows
The previous commit’s trick of creating a non-zero-size mapping for an empty
file does not work on Windows like it does with `mmap` on Unix.
It cause an `IoError` with code 1006 and message `The volume for a file has
been externally altered so that the opened file is no longer valid.`

https://github.com/RazrFalcon/memmap2-rs/pull/25/checks?check_run_id=3620122580#step:4:46

Instead, in that case skip creating a mapping entirely.
We don’t need one for `Deref` to return a valid zero-size slice.