Always use `-latest` runner images in CI
The `macos-10.15` image is deprecated and will be removed by 2022-08-30 [1]. It would be possible
to pin CI to `macos-11` or `macos-12`, but relying on `-latest` is unlikely to cause problems and
avoids having to update the workflow for this again.
While at it, also switch to `windows-latest` and `ubuntu-latest`.

Note that `-latest` are the latest stable images that GitHub provides, and might not be the most
recent version of the operating system available from the operating system vendor.

As of this writing this effectively updates:
macos-10.15 -> macos-11
windows-2019 -> windows-2022
(ubuntu remains unchanged at 20.04)

[1] https://github.blog/changelog/2022-07-20-github-actions-the-macos-10-15-actions-runner-image-is-being-deprecated-and-will-be-removed-by-8-30-22/