Add a dependabot config
Dependabot will automatically file PRs for updatable dependencies. The cargo section watches
Cargo.toml. The github-actions section watches the workflow files in .github/workflows for possible
updates to any of the Actions depended upon.

Ignore Dependabot's branches on CI to avoid running the tests twice, once on the push to the branch
and once on the creation of the PR.