Add support for mremap on Linux
This commit adds adds bindings for mremap(2) on Linux. It introduces the
following new methods:
- Mmap::remap
- MmapMut::remap
- MmapRaw::remap
and the RemapOptions type as a builder for supported options.

The only flag supported in this commit is the MREMAP_MAYMOVE flag (as
RemapOptions::may_move). I have left out MREMAP_FIXED and
MREMAP_DONTUNMAP.

I have marked all the remap functions as unsafe since I think that being
able to resize past the end of a file allowing the program to generate
SIGBUS signals by accessing the mapped memory probably qualifies as
undefined behaviour in rust.