clippy suspicious_open_options
`create` does not truncate by default. If less data is written to the
file than it already constains, the remainder of the file will remain unchanged and the end of the file will contain old data.

`truncate(true)` ensures the file is entirely overwritten with new data.

Formatting for examples was not fixed. That will be dealt with later.